Unify advancements screen into single shared component for lobby and pause menu
- Merge Advancements.tsx and AdvancementsMenu.tsx into one AdvancementsMenu component
  with an `overlay` prop for lobby usage (fullscreen overlay) vs pause menu (embedded)
- Fix tab button sizes: use flex: 1 so all category tabs are equal width
- Fix category switching: remove AnimatePresence mode="wait" that caused content
  to flash/disappear when clicking category tabs (e.g., general not showing data)
- Add i18n translation support (was missing in pause menu version)
- Delete redundant Advancements.tsx and Advancements.module.css
