+++ src/components/rhythmia/AdvancementsMenu.module.css
@@ -105,12 +105,13 @@
 }
 
 .tab {
+  flex: 1;
   display: flex;
   flex-direction: column;
   align-items: center;
   gap: 2px;
   padding: 6px 10px;
-  min-width: 56px;
+  min-width: 0;
   font-size: 0.55rem;
   background: rgba(50, 45, 35, 0.6);
   border: 1px solid rgba(80, 70, 50, 0.3);
@@ -294,7 +295,6 @@
   }
 
   .tab {
-    min-width: 44px;
     padding: 5px 6px;
   }
